Remove all spaces/format chars from data files
The use of spaces and/or other punctuation in filenames has been
problematic for years. Even though modern OSes (usually) handle it
somewhat OK, that can't be assumed. Older OSes (especially big *NIXes)
could be especially problematic.

This renames all data files in the thermo/ directory to use underscores
instead of spaces, removes (), and changes "_-_" to a single underscore.
References to the files are also updated.

I ran the test suite and the results are the same pre/post. (Some of the
    tests failed on a clean copy of master -- is the test suite 100%?)